+++ NEWS.md
@@ -2,6 +2,7 @@
 
 BREAKING CHANGES :
 * Add UTF-8 encoding argument in `.getJobs()`
+* Unit tests no longer call the study in the antaresRead package for versions > 8.0.0
 
 BUGFIXES :
 * `createArea()`/`editArea()` : in API mode, split data in nodalOptimization argument to write it in the expected files